fix gpu particle system and put ship thruster options in state
