+++ app/styles/app.css
@@ -105,7 +105,7 @@ a {
 
 #ui .scene-messages {
   width: 400px;
-  height: 200px;
+  height: 250px;
   position: absolute;
   bottom: 25px;
   left: 25px;
