+++ packages/core/sim/scene.js
@@ -1,6 +1,3 @@
-import { EffectComposer, BloomPass, RenderPass } from 'postprocessing'
-
-import * as weapons from '-/player/weapons'
 import { updateSystemCPU, updateSystemGPU } from './system'
 
 /**
@@ -18,27 +15,6 @@ const squareGrid = () => {
   const size = 100000000
   const divisions = 1000
   const gridHelper1 = new THREE.GridHelper(size, divisions, 0xffffff, 0xfffff)
-}
-
-/**
- * add post processing effects, e.g., bloom filter
- */
-const createPostprocessing = ({ renderer, scene, camera }) => {
-  const composer = new EffectComposer(renderer)
-  composer.addPass(new RenderPass(scene, camera))
-
-  const bloomPass = new BloomPass({
-    resolutionScale: 0.05,
-    kernelSize: 3.0,
-    intensity: 0.3,
-    distinction: 1
-  })
-  bloomPass.renderToScreen = true
-  bloomPass.combineMaterial.defines.SCREEN_MODE = '1'
-  bloomPass.combineMaterial.needsUpdate = true
-  composer.addPass(bloomPass)
-
-  return composer
 }
 
 /**
@@ -94,36 +70,6 @@ const animate = ({
   }
 
   // updateOimoPhysics()
-
-  composer.render(delta)
 }
 
-const createRenderer = () => {
-  const renderer = new THREE.WebGLRenderer({
-    antialias: true,
-    logarithmicDepthBuffer: true
-  })
-  renderer.setPixelRatio(window.devicePixelRatio)
-  renderer.setSize(window.innerWidth, window.innerHeight)
-  return renderer
-}
-
-const createCamera = options => {
-  const camera = new THREE.PerspectiveCamera(
-    options.fov,
-    window.innerWidth / window.innerHeight,
-    options.nearClip,
-    options.farClip
-  )
-  return camera
-}
-
-export {
-  addLights,
-  animate,
-  createCamera,
-  createPostprocessing,
-  createRenderer,
-  createUniverse,
-  squareGrid
-}
+export { addLights, animate, createUniverse, squareGrid }
