+++ packages/core/viewer.js
@@ -1,11 +1,7 @@
 import init from './sim'
 
-const createViewer = (rootId, animateCallbackHelpers, options) => {
-  const root = document.getElementById(rootId)
-  if (!root) {
-    console.error('invalid root element')
-  }
-  return init(root, animateCallbackHelpers, options)
+const createViewer = (scene, options) => {
+  return init(scene, options)
 }
 
 export { createViewer }
